handlers.c: fix warning about incorrect format string

+++ handlers.c
@@ -781,8 +781,8 @@ void handle_ext_l2cachefeat(struct cpu_regs_t *regs, __unused struct cpuid_state
 
 		l3_cache = (struct l3_cache_t *)&regs->edx;
 		if (l3_cache->size) {
-			size_t size = l3_cache->size * 512;
-			printf("L3 cache: %d%cB, %s, %d lines per tag, %d byte line size\n",
+			uint32_t size = l3_cache->size * 512;
+			printf("L3 cache: %u%cB, %s, %d lines per tag, %d byte line size\n",
 			       size > 1024 ? size / 1024 : size,
 			       size > 1024 ? 'M' : 'K',
 			       assoc[l3_cache->assoc] ? assoc[l3_cache->assoc] : "unknown associativity",
